pyproject.toml
[tool.poetry]
name = "work-register"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "work_register"}]

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.1.7"
djangorestframework = "^3.14.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.2.2"
pytest-django = "^4.5.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
